+++ api/src/config.js
@@ -1,4 +1,4 @@
-const { version, mobileAppVersion, mobileAppBuildNumber } = require("../package.json");
+const { version, mobileAppVersion, mobileAppAndroidBuildNumber, mobileAppIosBuildNumber } = require("../package.json");
 
 const PORT = process.env.PORT || 3000;
 const ENVIRONMENT = process.env.ENVIRONMENT || process.env.NODE_ENV || "development";
@@ -7,7 +7,8 @@ const SENTRY_KEY = process.env.SENTRY_KEY || "https://b43d73353b7b48b8857deb69bc
 
 const VERSION = version;
 const MOBILE_VERSION = mobileAppVersion;
-const MOBILE_BUILD_NUMBER = mobileAppBuildNumber;
+const MOBILE_ANDROID_BUILD_NUMBER = mobileAppAndroidBuildNumber;
+const MOBILE_IOS_BUILD_NUMBER = mobileAppIosBuildNumber;
 const MINIMUM_MOBILE_BUILD_NUMBER = 116; // manually change this number to force update app
 
 if (process.env.NODE_ENV === "development") {
@@ -17,7 +18,8 @@ if (process.env.NODE_ENV === "development") {
     SENTRY_KEY,
     VERSION,
     MOBILE_VERSION,
-    MOBILE_BUILD_NUMBER,
+    MOBILE_ANDROID_BUILD_NUMBER,
+    MOBILE_IOS_BUILD_NUMBER,
     MINIMUM_MOBILE_BUILD_NUMBER,
   });
 }
@@ -28,6 +30,7 @@ module.exports = {
   SENTRY_KEY,
   VERSION,
   MOBILE_VERSION,
-  MOBILE_BUILD_NUMBER,
+  MOBILE_ANDROID_BUILD_NUMBER,
+  MOBILE_IOS_BUILD_NUMBER,
   MINIMUM_MOBILE_BUILD_NUMBER,
 };