+++ app/src/context/needUpdate.js
@@ -15,8 +15,8 @@ export const NeedUpdateContextProvider = ({ children }) => {
       if (response.ok && BUILD_NUMBER < response.data.MOBILE_BUILD_NUMBER) {
         setNeedUpdate(true);
         Alert.alert(
-          `La nouvelle version ${response.data.MOBILE_VERSION}(${response.data.MOBILE_BUILD_NUMBER}) de Jardin Mental est disponible !`,
-          `Vous avez la version ${VERSION}(${BUILD_NUMBER}) actuellement sur votre téléphone`,
+          `Une nouvelle version de Jardin Mental est disponible !`,
+          ` ${response.data.MOBILE_VERSION}.${response.data.MOBILE_BUILD_NUMBER}`,
           [
             {
               text: "Télécharger",